intg: fix schema def

+++ entc/integration/privacy/ent/ent.go
@@ -255,8 +255,8 @@ func insertLastID(ctx context.Context, tx dialect.Tx, insert *sql.InsertBuilder)
 }
 
 // keys returns the keys/ids from the edge map.
-func keys(m map[uint64]struct{}) []uint64 {
-	s := make([]uint64, 0, len(m))
+func keys(m map[int]struct{}) []int {
+	s := make([]int, 0, len(m))
 	for id := range m {
 		s = append(s, id)
 	}
