+++ entc/integration/privacy/ent/planet/where.go
@@ -13,28 +13,28 @@ import (
 )
 
 // ID filters vertices based on their identifier.
-func ID(id uint64) predicate.Planet {
+func ID(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id uint64) predicate.Planet {
+func IDEQ(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id uint64) predicate.Planet {
+func IDNEQ(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldID), id))
 	})
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...uint64) predicate.Planet {
+func IDIn(ids ...int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		// if not arguments were provided, append the FALSE constants,
 		// since we can't apply "IN ()". This will make this predicate falsy.
@@ -51,7 +51,7 @@ func IDIn(ids ...uint64) predicate.Planet {
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...uint64) predicate.Planet {
+func IDNotIn(ids ...int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		// if not arguments were provided, append the FALSE constants,
 		// since we can't apply "IN ()". This will make this predicate falsy.
@@ -68,28 +68,28 @@ func IDNotIn(ids ...uint64) predicate.Planet {
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id uint64) predicate.Planet {
+func IDGT(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldID), id))
 	})
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id uint64) predicate.Planet {
+func IDGTE(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldID), id))
 	})
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id uint64) predicate.Planet {
+func IDLT(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldID), id))
 	})
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id uint64) predicate.Planet {
+func IDLTE(id int) predicate.Planet {
 	return predicate.Planet(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldID), id))
 	})
